refactor: move StandardPrinter to impl
